Clang-format, fixed include order.
Separate mip tests local (core) from remote (fetched) source files.
Activate s390 test build.
